test for times as well
