+++ tap_mysql/sync_strategies/common.py
@@ -7,12 +7,37 @@ import singer
 import time
 import tzlocal
 import pytz
+import pymysql
 
 import singer.metrics as metrics
 from singer import metadata
 from singer import utils
 
 LOGGER = singer.get_logger()
+
+
+# NB: Upgrading pymysql from 0.7.11 --> 0.9.3 had the undocumented change
+# to how `0000-00-00 00:00:00` date/time types are returned. In 0.7.11,
+# they are returned as NULL, and in 0.9.3, they are returned as the string
+# `0000-00-00 00:00:00`. To maintain backwards-compatability, we are
+# monkey patching the functions continue returning None
+original_convert_datetime = pymysql.converters.convert_datetime
+original_convert_date = pymysql.converters.convert_date
+
+def monkey_patch_datetime(datetime_str):
+    value = original_convert_datetime(datetime_str)
+    if datetime_str == value:
+        return None
+    return value
+
+def monkey_patch_date(date_str):
+    value = original_convert_date(date_str)
+    if date_str == value:
+        return None
+    return value
+
+pymysql.converters.convert_datetime = monkey_patch_datetime
+pymysql.converters.convert_date = monkey_patch_date
 
 def escape(string):
     if '`' in string:
